[Internal/CRI] Fix MXFP4 Group GeMM kernel nullptr issue; Re-enable CI tests (#110)
* Fix MXFP4 Group GeMM kernel nullptr issue; Re-enable CI tests

* Update blockwise_moe_kernel.cpp

* Fix MXFP4/MXFP8 blockwise MoE grouped GEMM CI failures

Set ElementC=void on the MXFP8 generic-group epilogue to elide the
C-source load (fixes XE_STORE_2D misalignment assert); keep the MXFP4
legacy path as-is since that specialization doesn't compile with void.

* Split blockwise MoE kernel into per-format TUs (MXFP4/MXFP8)

MXFP4 needs CUTE_ENABLE_XE_BLOCK_2D_ASSERT as a compiler barrier around
block-2D payload setup, but the same define makes MXFP8's D-store fail
an x%4 alignment assert that the hardware silently handles. Split the
combined TU so each path controls the macro independently:

  - blockwise_moe_runner.hpp  shared header-only runner template
  - blockwise_moe_mxfp4.cpp   defines CUTE_ENABLE_XE_BLOCK_2D_ASSERT,
                              hosts MXFP4Types (legacy IntelXeXMX16Group
                              epilogue, ElementC=ElementAccumulator)
  - blockwise_moe_mxfp8.cpp   no define, hosts MXFP8Types (generic
                              IntelXeGenericGroup epilogue,
ElementC=void)

Also adds moe_group_gemm_helper.hpp.

Build picks up the new sources automatically via the existing
GLOB_RECURSE in src/CMakeLists.txt; no build-system edits required.

Verified locally:
  test_mxfp4_blockwise_moe::test_kernel_vs_reference[2-64-64-64]    PASS
  test_mxfp8_blockwise_moe::test_kernel_vs_reference[2-128-128-128] PASS
  test_mxfp8_blockwise_moe::test_single_expert                      PASS


---------
